requirements_old.txt
alembic==1.7.1
amqp==5.0.6
aniso8601==9.0.1
astroid==2.2.5
atomicwrites==1.4.0
attrs==21.2.0
awscli==1.21.9
Babel==2.9.0
billiard==3.6.4.0
botocore==1.22.9
celery==5.2.1
cffi==1.14.6
click==8.0.1
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
cryptography==3.4.8
docutils==0.15.2
ecdsa==0.17.0
Flask==1.1.2
Flask-Babel==2.0.0
Flask-Cors==3.0.10
Flask-Migrate==1.8.0
Flask-Moment==0.11.0
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-Script-Extras==0.1.4
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
future==0.18.2
greenlet==1.1.1
gunicorn==20.1.0
iniconfig==1.1.1
isort==4.3.18
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
jwt==1.2.0
kombu==5.2.2
lazy-object-proxy==1.4.0
Mako==1.1.5
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.2
pluggy==0.13.1
postgres==3.0.0
prompt-toolkit==3.0.23
psycopg2==2.9.1
psycopg2-binary==2.9.1
psycopg2-pool==1.1
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
pycryptodome==3.11.0
PyJWT==1.7.1
pylint==2.3.1
pyparsing==2.4.7
pytest==6.2.2
python-dateutil==2.6.0
python-jose==3.3.0
pytz==2019.1
PyYAML==5.4.1
rsa==4.7.2
s3transfer==0.5.0
six==1.12.0
SQLAlchemy==1.4.23
toml==0.10.2
typed-ast==1.4.2
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.11.1
WTForms==2.3.3
